refactored code base, made classes to manage breakout boards and handle state determination

+++ carm-electronics/flight-computer/bitpack.h
@@ -1,23 +1,23 @@
 #ifndef BITPACK_H
 #define BITPACK_H
 
-static uint32_t
-shift_left(uint32_t word, unsigned bits);
+static uint64_t
+shift_left(uint64_t word, unsigned bits);
 
-static uint32_t
-shift_right_logical(uint32_t word, unsigned bits);
+static uint64_t
+shift_right_logical(uint64_t word, unsigned bits);
 
-static int32_t
-shift_right_arith(uint32_t word, unsigned bits);
+static int64_t
+shift_right_arith(uint64_t word, unsigned bits);
 
-bool Bitpack_fitsu(uint32_t n, unsigned width);
+bool Bitpack_fitsu(uint64_t n, unsigned width);
 
-bool Bitpack_fitss(int32_t n, unsigned width);
+bool Bitpack_fitss(int64_t n, unsigned width);
 
-uint32_t
-Bitpack_getu(uint32_t word, unsigned width, unsigned lsb);
+uint64_t
+Bitpack_getu(uint64_t word, unsigned width, unsigned lsb);
 
-uint32_t
-Bitpack_newu(uint32_t word, unsigned width, unsigned lsb, uint32_t value);
+uint64_t
+Bitpack_newu(uint64_t word, unsigned width, unsigned lsb, uint64_t value);
 
 #endif
